requirements-additional.txt
# Additional requirements for enterprise features
# Add these to your main requirements.txt

# Email Service Dependencies
aiosmtplib==3.0.1
email-validator==2.1.0
bleach==6.1.0
jinja2==3.1.3

# Monitoring & Telemetry Dependencies
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-instrumentation-httpx==0.45b0
opentelemetry-instrumentation-asyncpg==0.45b0
opentelemetry-instrumentation-celery==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-propagator-b3==1.24.0

# Additional Security Dependencies
python-multipart==0.0.9
magic-python==0.4.27

# System Monitoring
psutil==5.9.8

# AWS Dependencies (optional, for AWS SES)
boto3==1.34.28

# SendGrid (optional)
sendgrid==6.11.0

# Structured Logging (if not already included)
structlog==24.1.0
python-json-logger==2.0.7